+++ sqlalchemy_redshift/dialect.py
@@ -13,7 +13,7 @@ from sqlalchemy.sql.expression import (
 )
 from sqlalchemy.types import VARCHAR, NullType
 
-from .commands import CopyCommand, UnloadFromSelect
+from .commands import CopyCommand, UnloadFromSelect, CreateTableAs
 from .compat import string_types
 
 try:
@@ -27,7 +27,8 @@ else:
     class RedshiftImpl(postgresql.PostgresqlImpl):
         __dialect__ = 'redshift'
 
-__all__ = ['CopyCommand', 'UnloadFromSelect', 'RedshiftDialect']
+__all__ = ['CopyCommand', 'UnloadFromSelect',
+           'RedshiftDialect', 'CreateTableAs']
 
 
 # Regex for parsing and identity constraint out of adsrc, e.g.: